workers/Worker_ErrorEvent_filename.htm

<!DOCTYPE html>
<title> AbstractWorker ErrorEvent.filename </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onerror = t.step_func_done(function(e) {
    assert_equals(e.filename, "Error Message");
  });
  worker.postMessage("Error Message");
});
</script>
